% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisper.R
\name{predict.whisper}
\alias{predict.whisper}
\title{Transcribe audio files using a Whisper model}
\usage{
\method{predict}{whisper}(object, newdata, language = "en", ...)
}
\arguments{
\item{object}{a whisper object}

\item{newdata}{the path to a 16-bit .wav file}

\item{language}{the language of the audio. Defaults to 'en'}

\item{...}{further arguments, not used currently}
}
\value{
a list with the following elements:
\itemize{
\item{segments: the number of audio segments}
\item{data: a data.frame with the transcription with columns text, from and to}
\item{params: a list with parameters}
}
}
\description{
Automatic Speech Recognition using Whisper on 16-bit WAV files
}
\seealso{
\code{\link{whisper}}
}
