% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vad-channel.R
\name{predict.whisper_transcription}
\alias{predict.whisper_transcription}
\title{Predict to which channel a transcription section belongs}
\usage{
\method{predict}{whisper_transcription}(object, vad, type = "channel", threshold = 0, ...)
}
\arguments{
\item{object}{an object of class \code{whisper_transcription} as returned by \code{\link{predict.whisper}}}

\item{vad}{an object of class \code{webrtc-gmm-bychannel} as returned by function \code{VAD_channel} from R package \code{audio.vadwebrtc} with information of the detected voice in at least channels 1 and 2.
ar a list with element vad_segments containing a data.frame with columns channel, start, end and has_voice with information at which second
there was a voice in the audio}

\item{type}{character string with currently only possible value: 'channel' which does a 2-speaker channel assignment}

\item{threshold}{numeric in 0-1 range indicating if the difference between the probability that the segment was from the left channel 1 or the right channel 2 is smaller than this amount, the column \code{channel} will be set to 'both'. Defaults to 0.}

\item{...}{not used}
}
\value{
an object of class \code{whisper_transcription} as documented in \code{\link{predict.whisper}} 
where element \code{data} contains the following extra columns indicating which channel the transcription is probably from
\itemize{
\item{channel: either 'left', 'right' or 'both' indicating the transcription segment was either from the left channel (1), the right channel (2) or probably from both as identified by the Voice Activity Detecion}
\item{channel_probability: a number between 0 and 1 indicating for that specific segment the ratio of the amount of voiced seconds in the most probably channel to 
the sum of the amount of voiced seconds in the left + the right channel}
\item{duration: how long (in seconds) the from-to segment is}
\item{duration_voiced_left: how many seconds there was a voiced signal on the left channel (channel 1) as identified by \code{vad}}
\item{duration_voiced_right: how many seconds there was a voiced signal on the right channel (channel 2) as identified by \code{vad}}
}
}
\description{
Audio files containing 2 channels which were transcribed with \code{\link{predict.whisper}}, 
you can use the results of a Voice Activity Detection by channel (either with R packages 
\code{audio.vadwebrtc} or \code{audio.vadsilero}) to assign the text segments to each of the channels.\cr
This is done by looking for each text segment how many seconds overlap there is with the voiced sections which are identified
by the Voice Activity Detection.
}
\examples{
library(audio.whisper)
model <- whisper("tiny")
audio <- system.file(package = "audio.whisper", "samples", "stereo.wav")
trans <- predict(model, audio, language = "es")
\dontrun{
library(audio.vadwebrtc)
vad   <- VAD_channel(audio, channels = "all", mode = "veryaggressive", milliseconds = 30)
}
vad   <- list(vad_segments = rbind(
  data.frame(channel = 1, start = c(0, 5, 15, 22), end = c(5, 9, 18, 23), has_voice = TRUE),
  data.frame(channel = 2, start = c(2, 9.5, 19, 22), end = c(2.5, 13.5, 21, 23), has_voice = TRUE)))
out <- predict(trans, vad, type = "channel", threshold = 0)
out$data
}
\seealso{
\code{\link{predict.whisper}}
}
