% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vad.R
\name{vad}
\alias{vad}
\title{Voice Activity Detection using Silero}
\usage{
vad(
  path = system.file(package = "audio.whisper", "samples", "jfk.wav"),
  vad_model = system.file(package = "audio.whisper", "silero", "ggml-silero-v5.1.2.bin"),
  ...
)
}
\arguments{
\item{path}{TODO}

\item{vad_model}{TODO}

\item{...}{passed on to the C++ silero_vad function}
}
\value{
TODO
}
\description{
Voice Activity Detection using Silero
}
\examples{
audio <- system.file(package = "audio.whisper", "samples", "jfk.wav")
voice <- vad(audio)
}
\seealso{
\code{\link{predict.whisper}}
}
