% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisper.R
\name{whisper_benchmark}
\alias{whisper_benchmark}
\title{Benchmark a Whisper model}
\usage{
whisper_benchmark(
  object = whisper(system.file(package = "audio.whisper", "models",
    "for-tests-ggml-tiny.bin")),
  threads = 1
)
}
\arguments{
\item{object}{a whisper object}

\item{threads}{the number of threads to use, defaults to 1}
}
\value{
invisible()
}
\description{
Benchmark a Whisper model to see how good it runs on your architecture by printing it's performance on 
fake data. \url{https://github.com/ggerganov/whisper.cpp/issues/89}
}
\examples{
\dontrun{ 
model <- whisper("tiny")
whisper_benchmark(model)
}
}
\seealso{
\code{\link{whisper}}
}
