% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisper.R
\name{whisper_languages}
\alias{whisper_languages}
\title{Get the language capabilities of Whisper}
\usage{
whisper_languages()
}
\value{
a data.frame with columns id, language and language_label showing the languages
}
\description{
Extract the list of languages a multilingual whisper model is able to handle
}
\examples{
x <- whisper_languages()
x
}
