% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisper.R
\name{predict.whisper}
\alias{predict.whisper}
\title{Transcribe audio files using a Whisper model}
\usage{
\method{predict}{whisper}(
  object,
  newdata,
  type = c("transcribe", "translate"),
  language = "auto",
  trim = FALSE,
  ...
)
}
\arguments{
\item{object}{a whisper object}

\item{newdata}{the path to a 16-bit .wav file}

\item{type}{character string with the type of prediction, can either be 'transcribe' or 'translate'}

\item{language}{the language of the audio. Defaults to 'auto'}

\item{trim}{logical indicating to trim leading/trailing white space from the transcription using \code{\link{trimws}}. Defaults to \code{FALSE}.}

\item{...}{further arguments, directly passed on to the C++ function, for expert usage only}
}
\value{
an object of class \code{whisper_transcription} which is a list with the following elements:
\itemize{
\item{n_segments: the number of audio segments}
\item{data: a data.frame with the transcription with columns segment, text, from and to}
\item{tokens: a data.frame with the transcription tokens with columns segment, token_id, token, token_prob indicating the token probability given the context}
\item{params: a list with parameters used for inference}
\item{timing: a list with elements start, end and duration indicating how long it took to do the transcription}
}
}
\description{
Automatic Speech Recognition using Whisper on 16-bit WAV files
}
\examples{
\donttest{ 
model <- whisper("tiny")
audio <- system.file(package = "audio.whisper", "samples", "jfk.wav")
trans <- predict(model, newdata = audio)
trans <- predict(model, newdata = audio, language = "en")
trans <- predict(model, newdata = audio, language = "en", token_timestamps = TRUE)

audio <- system.file(package = "audio.whisper", "samples", "proficiat.wav")
trans <- predict(model, newdata = audio, language = "nl", type = "translate")
}

## Predict using a quantised model
audio <- system.file(package = "audio.whisper", "samples", "jfk.wav")
path  <- system.file(package = "audio.whisper", "repo", "ggml-tiny.en-q5_1.bin")
model <- whisper(path)
trans <- predict(model, newdata = audio, language = "en")
trans <- predict(model, newdata = audio, language = "en", token_timestamps = TRUE)
}
\seealso{
\code{\link{whisper}}, \code{\link{whisper_languages}}
}
