% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisper.R
\name{predict.whisper}
\alias{predict.whisper}
\title{Transcribe audio files using a Whisper model}
\usage{
\method{predict}{whisper}(object, newdata, language = "en", ...)
}
\arguments{
\item{object}{a whisper object}

\item{newdata}{the path to a 16-bit .wav file}

\item{language}{the language of the audio. Defaults to 'en'}

\item{...}{further arguments, for expert usage only}
}
\value{
a list with the following elements:
\itemize{
\item{n_segments: the number of audio segments}
\item{segments: a data.frame with the transcription with columns segment, text, from and to}
\item{tokens: a data.frame with the transcription tokens with columns segment, token, token_prob indicating the token probability given the context}
\item{params: a list with parameters used for inference}
}
}
\description{
Automatic Speech Recognition using Whisper on 16-bit WAV files
}
\examples{
\dontrun{ 
model <- whisper("tiny")
audio <- system.file(package = "audio.whisper", "samples", "jfk.wav")
trans <- predict(model, newdata = audio)
trans <- predict(model, newdata = audio, token_timestamps = TRUE)
}
}
\seealso{
\code{\link{whisper}}
}
