% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisper.R
\name{whisper}
\alias{whisper}
\title{Automatic Speech Recognition using Whisper}
\usage{
whisper(x, ...)
}
\arguments{
\item{x}{the path to a model, an object returned by \code{\link{whisper_download_model}} or a character string with 
the name of the model which can be passed on to \code{\link{whisper_download_model}}}

\item{...}{further arguments, not used currently}
}
\value{
a list with the following elements: TODO
}
\description{
Automatic Speech Recognition using Whisper on 16-bit WAV files
}
\examples{
\dontrun{ 
model <- whisper("tiny")
trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"))
trans
model <- whisper("medium")
trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"))

## Or download the model explicitely
path  <- whisper_download_model("tiny")
model <- whisper(path)
trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"))
}

\dontshow{
## Or provide the path to the model
path  <- system.file(package = "audio.whisper", "models", "for-tests-ggml-tiny.bin")
model <- whisper(path)
trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"))
}
}
