% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisper.R
\name{whisper}
\alias{whisper}
\title{Automatic Speech Recognition using Whisper}
\usage{
whisper(x, overwrite = FALSE, model_dir = getwd(), ...)
}
\arguments{
\item{x}{the path to a model, an object returned by \code{\link{whisper_download_model}} or a character string with 
the name of the model which can be passed on to \code{\link{whisper_download_model}}}

\item{overwrite}{logical indicating to overwrite the file if the file was already downloaded, passed on to \code{\link{whisper_download_model}}. Defaults to \code{FALSE}.}

\item{model_dir}{a path where the model will be downloaded to, passed on to \code{\link{whisper_download_model}}. Defaults to the current working directory.}

\item{...}{further arguments, passed on to the internal C++ function \code{whisper_load_model}}
}
\value{
an object of class \code{whisper} which is list with the following elements: 
\itemize{
\item{file: path to the model}
\item{model: an Rcpp pointer to the loaded Whisper model}
}
}
\description{
Automatic Speech Recognition using Whisper on 16-bit WAV files
}
\examples{
\dontrun{ 
model <- whisper("tiny")
trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"))
trans
model <- whisper("base")
trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"))
trans
model <- whisper("small")
trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"))
trans
model <- whisper("medium")
trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"))
trans
model <- whisper("large-v1")
trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"))
trans

## Or download the model explicitely
path  <- whisper_download_model("tiny")
model <- whisper(path)
trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"))
}

\dontshow{
## Or provide the path to the model
path  <- system.file(package = "audio.whisper", "models", "for-tests-ggml-tiny.bin")
model <- whisper(path)
trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"), 
                 language = "en", duration = 1000)
path  <- system.file(package = "audio.whisper", "models", "for-tests-ggml-tiny.en.bin")
model <- whisper(path)
trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"), 
                 language = "en", duration = 1000)
}
}
